Create board class and modify node class
I wrote a board class which stores all the nodes in a has where the key
is equal to the node's location on the field
it also contains several methods to display the chess board and al the
chess pieces inside the nodes.

I had to modify the node class as the piece variable was immutable.