Create chess pieces
Each chess piece is represented by it's own class containing the color,
utf-chess-piece-symbol, location on the board as well as the base moves.
There is also a variable for an array to store all the possible moves
from the current location.
There are also tests to make sure the symbols are set right and for the
pawn to have the correct moves.