Create board setup module and adapt Node class
I wrote several methods to create the required chess pieces and place
them on the right fields of the board represented by a node.
I adapted the Node class and added another variable containing the
symbol of the chess piece to make displaying the board easier in case
there is no piece on the field.