Create method to find all possible moves
I wrote a method that fills the array of every possible moves for a
piece according to its available moves. I added an argument to switch
for it to work for single-move pieces like king, knight and pawn as well
as pieces that can move several squares in a row.
I also adapted the pawn move list as I got the axis wrong and no longer
store a complete node as location but just the node's location.