Build Game class for game logic
This class contains all the logic to play chess. In this first draft the
player can move around the chess pieces, but it still lacks the logic
for check or check mate.

I had to adapt the pawn moves as well as the Node class, because the
possible moves were created from the piece's location so it had to be
mutable.