Create check mate logic
I added an instance variable hash to the King class which stores the
every piece giving check as well as all the lines through which it
checks the king.
Should a king be in check after the opponents move, a method checks
if the king has a valid move. If it has not, there is a method to check
whether a piece can intercept the check and if not, check mate is called